+++ src/core/crypto/hs_ntor.h
@@ -62,7 +62,7 @@ int hs_ntor_service_get_introduce1_keys_multi(
             const struct ed25519_public_key_t *intro_auth_pubkey,
             const struct curve25519_keypair_t *intro_enc_keypair,
             const struct curve25519_public_key_t *client_ephemeral_enc_pubkey,
-            int n_subcredentials,
+            size_t n_subcredentials,
             const hs_subcredential_t *subcredentials,
             hs_ntor_intro_cell_keys_t *hs_ntor_intro_cell_keys_out);
